Use direct reference to a matcher from the closure, because it's safe.
While passing `feed` onto each goroutine stack makes sense because this
variable is reused during each for-iteration, doing the same for matcher
is redundant. This variable is created on each for-iteration, and thus
is safe: https://play.golang.org/p/83Y4sGhFpd3

In the Go play example above vet complaints only about case when feed is
captured, closing over the matcher is fine.